Xtend_WebRequest geupdated
Changed; WebRequest.PathInfo As Dictionary. Now uses path based separation /key1/value1/key2/value2/key3 in this "key1" will have a value of "value1" and "key3" will have a value of "".
Added; WebRequest.RedirectPermanently(ToURL As String) sends a 301 Moved Permanently response with Location header.
Added; WebRequest.RedirectTemporary(ToURL As String) sends a 302 Moved Temporary response with Location header.
Added WebRequest.Host  As String to quickly get the host (or ip) where the request was originated to. This can be "127.0.0.1" depending on the setup. For example if the request was handled by nginx then it could be "127.0.0.1"